requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.0.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
coverage==7.9.2
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.14
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
playwright==1.53.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-base-url==2.1.0
pytest-cov==6.2.1
pytest-playwright==0.7.0
python-decouple==3.8
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
python-slugify==8.0.4
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
text-unidecode==1.3
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.35.0
